+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased
 
+## V7.2.0 - 2025-01-30
+
 - Allow the `union` operation between `FiniteDatetimeRange` and `HalfFiniteDatetimeRange` 
   (and other `DatetimeRange`s) [#189](https://github.com/octoenergy/xocto/pull/189/).